+++ noxfile.py
@@ -49,9 +49,7 @@ def djlint(session):
 
 
 DJANGO_PYTHON_REQ = {
-    "4.2.20": ("3.8", "3.9", "3.10", "3.11", "3.12"),
-    # 5.0 is EOL
-    # "5.0": ("3.10", "3.11", "3.12"),
+    "4.2.20": ("3.10", "3.11", "3.12"),
     "5.1": ("3.10", "3.11", "3.12", "3.13"),
     "5.2": ("3.10", "3.11", "3.12", "3.13", "3.14"),
     "6.0": ("3.12", "3.13", "3.14"),
@@ -59,7 +57,7 @@ DJANGO_PYTHON_REQ = {
 DJANGO_LTS = "5.2"
 
 
-@nox.session(python=["3.8", "3.9", "3.10", "3.11", "3.12", "3.13", "3.14"])
+@nox.session(python=["3.10", "3.11", "3.12", "3.13", "3.14"])
 @nox.parametrize("django", list(DJANGO_PYTHON_REQ.keys()))
 @nox.parametrize(
     "project", ["regular", "headless_only", "account_only", "login_required_mw"]