+++ ChangeLog.rst
@@ -1,4 +1,4 @@
-65.15.0 (2025-03-09)
+65.15.1 (unreleased)
 ********************
 
 .. note::
@@ -6,6 +6,16 @@
     💙 **Is django-allauth's authentication the entrance to your business?**
     Please consider supporting its continued development by becoming a sponsor at
     https://allauth.org/sponsors/. Your support helps keep this project thriving!
+
+
+Note worthy changes
+-------------------
+
+- ...
+
+
+65.15.0 (2025-03-09)
+********************
 
 Note worthy changes
 -------------------